+++ internal/schema/config.go
@@ -1,9 +1,10 @@
 package schema
 
 type Config struct {
-	LogLevel        string `mapstructure:"log_level"`
-	Command         string `mapstructure:"tf_command"`
-	StacksDir       string `mapstructure:"stacks_dir"`
-	WorkDir         string `mapstructure:"work_dir"`
-	GenerateBackend bool   `mapstructure:"generate_backend"`
+	LogLevel        string            `mapstructure:"log_level"`
+	Command         string            `mapstructure:"tf_command"`
+	StacksDir       string            `mapstructure:"stacks_dir"`
+	WorkDir         string            `mapstructure:"work_dir"`
+	GenerateBackend bool              `mapstructure:"generate_backend"`
+	Env             map[string]string `mapstructure:"env"`
 }